* various glibc compat:    * gnu_get_libc_release() now returns "stable-musl-standalone" for better identification    * added various __isoc23_* C23 aliases to their respective equivalents (we don't perform any special handling)    * malloc: added mallinfo2 support (as in: symbols are there, but it's not actually implemented)    * malloc: added dummy/non-functional malloc hooks    * openat: implemented openat2 and added aliases    * time: made __strftime_l public    * auxv: made __getauxval() public    * stdlib: added strfromf128/strtof128 stubs    * epoll: added epoll_pwait2    * unistd: added __readlinkat_chk() and close_range()    * stdio: added unimplemented/aborting parse_printf_format()    * mbs*towcs: added __mbsnrtowcs_chk and __mbsrtowcs_chk    * wmem*: added __wmemcpy_chk and __wmemset_chk    * fcntl: added fcntl64 alias to fcntl    * pthread: made __pthread_* symbols public    * pthread: modify pthread_mutex_t, rwlock and barrier layout so that they are compatible with glibc (at least necessary for mutex, other are not exactly necessary)    * locale: __locale_struct is now ABI compatible with glibc + provides enough glibc compat for this to not cause any issue

+++ src/thread/pthread_mutex_trylock.c
@@ -16,7 +16,7 @@ int __pthread_mutex_trylock_owner(pthread_mutex_t *m)
 			goto success;
 		}
 		if ((type&3) == PTHREAD_MUTEX_RECURSIVE) {
-			if ((unsigned)m->_m_count >= INT_MAX) return EAGAIN;
+			if ((unsigned)m->_m_count >= SHRT_MAX) return EAGAIN;
 			m->_m_count++;
 			return 0;
 		}
